refactor(12618): small fixes
